Fix bug that disabled the bg video
On the first page visit, the videoPlayPreference local storage key is
set properly, but the `preference` var was set to `true` instead of
`'on'`. Now that it is set to `'on'`, the video should play on the
user's first time visiting the page.

Get rid of CNAME. That thing didn't work and I don't want github
sending me a page build warning email every time I push.